Remove extractFromFile
It seems `extractFromFile` was only used in tests but nowhere else.
`extractFromSourceFile` is the true public method.
